requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-rest-framework-simplejwt==5.2.2
django-cors-headers==4.0.0
django-phonenumber-field==7.2.0
phonenumbers==8.13.0
django-filter==23.2
django-redis==5.2.0
redis==4.5.5
python-decouple==3.8
celery==5.3.0
twilio==8.10.0
requests==2.31.0
pillow==10.0.0
django-storages==1.14.2
boto3==1.28.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
drf-spectacular==0.26.2
gunicorn==21.2.0
psycopg2-binary==2.9.7
pytest==7.4.0
pytest-django==4.5.2
whitenoise==6.5.0
